Ligthen ViewModel
